+++ source/fortran/input.f90
@@ -119,6 +119,7 @@ module process_input
   use times_variables
   use vacuum_variables
   use rebco_variables
+  use reinke_variables
 
   implicit none
 
@@ -129,7 +130,7 @@ module process_input
 #ifdef unit_test
   public :: parse_input_file
 #endif
-  public :: upper_case
+!  public :: upper_case
 
   integer, parameter :: maxlen = 300  !  maximum line length
   character(len=maxlen) :: line  !  current line of text from input file
@@ -554,6 +555,9 @@ contains
        case ('epbetmax')
           call parse_real_variable('epbetmax', epbetmax, 0.01D0, 10.0D0, &
                'Max epsilon*beta value')
+       case ('eped_sf')
+          call parse_real_variable('eped_sf', eped_sf, 0.0001D0, 2.0D0, &
+               'Scaling factor for EPED pedestal model')   
        case ('falpha')
           call parse_real_variable('falpha', falpha, 0.0D0, 1.0D0, &
                'Fraction of alpha power deposited to plasma')
@@ -949,6 +953,9 @@ contains
        case ('fradwall')
           call parse_real_variable('fradwall', fradwall, 0.001D0, 1.0D0, &
                'f-value for upper limit on radiation wall load')
+       case ('freinke')
+          call parse_real_variable('freinke', freinke, 0.001D0, 1.0D0, &
+               'f-value for upper limit on Reinke detachment criterion')
        case ('frminor')
           call parse_real_variable('frminor', frminor, 0.001D0, 10.0D0, &
                'F-value for minor radius limit')
@@ -995,6 +1002,9 @@ contains
        case ('fwalld')
           call parse_real_variable('fwalld', fwalld, 0.001D0, 10.0D0, &
                'F-value for wall load limit')
+       case ('fzactual')
+          call parse_real_variable('fzactual', fzactual, 0.0D0, 1.0D0, &
+               'fraction of specified impurity in SOL when constrained by Reinke criteria')
        case ('fzeffmax')
           call parse_real_variable('fzeffmax', fzeffmax, 0.001D0, 1.0D0, &
                'f-value for Zeff limit equation')
@@ -2760,7 +2770,7 @@ contains
                'Unplanned unavailability for vessel')
 
 
-          !  Sweep settings
+        !  Sweep settings
 
        case ('isweep')
           call parse_int_variable('isweep', isweep, 0, ipnscns, &
@@ -2771,6 +2781,19 @@ contains
        case ('sweep')
           call parse_real_array('sweep', sweep, isub1, ipnscns, &
                'Actual values to use in scan', icode)
+
+        case ('scan_dim')
+          call parse_int_variable('scan_dim', scan_dim, 1, 2, &
+               'Switch for 1-D or 2-D scan')
+        case ('isweep_2')
+          call parse_int_variable('isweep_2', isweep_2, 0, ipnscns, &
+               'Number of 2D scans to perform')
+        case ('nsweep_2')
+          call parse_int_variable('nsweep_2', nsweep_2, 1, ipnscnv, &
+               'Second variable used in 2D scan')
+        case ('sweep_2')
+          call parse_real_array('sweep_2', sweep_2, isub1, ipnscns, &
+               'Actual values to use in 2D scan', icode)
 
           !  Buildings settings
 
@@ -2918,8 +2941,14 @@ contains
           call parse_real_variable('outgasfactor', outgasfactor, 1.0D-6, 1.0D3, &
                'outgassing prefactor kw: outgassing rate at 1 s per unit area (Pa m s-1)')
 
+          ! Reinke criterion
+       case ('lhat')
+          call parse_real_variable('lhat', lhat, 1.0D0, 1.5D1, &
+               'connection length factor')
 
-
+       case ('impvardiv')
+          call parse_int_variable('impvardiv', impvardiv, 3, nimp, &
+               'Index of impurity to be iterated for Reike criterion')
 
 
           !  Stellarator settings
@@ -4780,67 +4809,6 @@ contains
 
   end subroutine check_range_real
 
-  ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-  subroutine upper_case(string,start,finish)
-
-    !+ad_name  upper_case
-    !+ad_summ  Routine that converts a (sub-)string to uppercase
-    !+ad_type  Subroutine
-    !+ad_auth  P J Knight, CCFE, Culham Science Centre
-    !+ad_cont  N/A
-    !+ad_args  string : input string   : character string of interest
-    !+ad_args  start  : optional input integer  : starting character for conversion
-    !+ad_args  finish : optional input integer  : final character for conversion
-    !+ad_desc  This routine converts the specified section of a string
-    !+ad_desc  to uppercase. By default, the whole string will be converted.
-    !+ad_prob  None
-    !+ad_call  None
-    !+ad_hist  05/01/04 PJK Initial F90 version
-    !+ad_hist  12/04/11 PJK Made start,finish arguments optional
-    !+ad_stat  Okay
-    !+ad_docs  None
-    !
-    ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-    implicit none
-
-    !  Arguments
-
-    character(len=*), intent(inout) :: string
-    integer, optional, intent(in) :: start,finish
-
-    !  Local variables
-
-    character(len=1) :: letter
-    character(len=27) :: upptab = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ_'
-    integer :: loop, i
-
-    integer :: first, last
-
-    ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-    if (present(start)) then
-       first = start
-    else
-       first = 1
-    end if
-
-    if (present(finish)) then
-       last = finish
-    else
-       last = len(string)
-    end if
-
-    if (first <= last) then
-       do loop = first,last
-          letter = string(loop:loop)
-          i = index('abcdefghijklmnopqrstuvwxyz_',letter)
-          if (i > 0) string(loop:loop) = upptab(i:i)
-       end do
-    end if
-
-  end subroutine upper_case
 
   ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!